# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Schemes Made Simple API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Scheme {
  id: number;
  name: string;
  description: string;
  category: string;
  state: string;
  ministry: string;
  launchDate: string;
  deadline?: string | null;
  benefits: string[];
  eligibility: string[];
  documents: string[];
  applicationUrl?: string | null;
  tags: string[];
}

export interface SchemeStep {
  stepNumber: number;
  title: string;
  description: string;
}

export interface SchemeFaq {
  question: string;
  answer: string;
}

export type SchemeDetail = Scheme & {
  steps: SchemeStep[];
  faqs: SchemeFaq[];
};

export interface SchemeListResponse {
  schemes: Scheme[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CategoriesResponse {
  categories: string[];
}

export interface ChatRequest {
  message: string;
  language?: string;
}

export interface ChatResponse {
  reply: string;
  sources: string[];
}

export interface TtsRequest {
  text: string;
  language?: string;
}

export interface TtsResponse {
  audioBase64: string;
  mimeType: string;
}

export type ListSchemesParams = {
  /**
   * Search query
   */
  search?: string;
  /**
   * Filter by category
   */
  category?: string;
  /**
   * Filter by state (use "central" for central schemes)
   */
  state?: string;
  page?: number;
  limit?: number;
};