# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Schemes Made Simple API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: schemes
    description: Government scheme operations
  - name: chat
    description: Chatbot operations
  - name: tts
    description: Text-to-speech operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /schemes:
    get:
      operationId: listSchemes
      tags: [schemes]
      summary: List all schemes
      description: Get a paginated list of government schemes with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search query
        - name: category
          in: query
          schema:
            type: string
          description: Filter by category
        - name: state
          in: query
          schema:
            type: string
          description: Filter by state (use "central" for central schemes)
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: List of schemes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemeListResponse"

  /schemes/{id}:
    get:
      operationId: getScheme
      tags: [schemes]
      summary: Get scheme details
      description: Get full details for a specific scheme
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scheme details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemeDetail"
        "404":
          description: Scheme not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /schemes/categories:
    get:
      operationId: listCategories
      tags: [schemes]
      summary: List all categories
      description: Get a list of all available scheme categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CategoriesResponse"

  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the chatbot
      description: Get a response from the legal awareness chatbot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chatbot response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /tts:
    post:
      operationId: textToSpeech
      tags: [tts]
      summary: Convert text to speech
      description: Convert text to speech audio (returns base64 audio)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TtsRequest"
      responses:
        "200":
          description: Audio data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TtsResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Scheme:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        state:
          type: string
        ministry:
          type: string
        launchDate:
          type: string
        deadline:
          type: string
          nullable: true
        benefits:
          type: array
          items:
            type: string
        eligibility:
          type: array
          items:
            type: string
        documents:
          type: array
          items:
            type: string
        applicationUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - description
        - category
        - state
        - ministry
        - launchDate
        - benefits
        - eligibility
        - documents
        - tags

    SchemeDetail:
      allOf:
        - $ref: "#/components/schemas/Scheme"
        - type: object
          properties:
            steps:
              type: array
              items:
                $ref: "#/components/schemas/SchemeStep"
            faqs:
              type: array
              items:
                $ref: "#/components/schemas/SchemeFaq"
          required:
            - steps
            - faqs

    SchemeStep:
      type: object
      properties:
        stepNumber:
          type: integer
        title:
          type: string
        description:
          type: string
      required:
        - stepNumber
        - title
        - description

    SchemeFaq:
      type: object
      properties:
        question:
          type: string
        answer:
          type: string
      required:
        - question
        - answer

    SchemeListResponse:
      type: object
      properties:
        schemes:
          type: array
          items:
            $ref: "#/components/schemas/Scheme"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - schemes
        - total
        - page
        - totalPages

    CategoriesResponse:
      type: object
      properties:
        categories:
          type: array
          items:
            type: string
      required:
        - categories

    ChatRequest:
      type: object
      properties:
        message:
          type: string
        language:
          type: string
          default: en
      required:
        - message

    ChatResponse:
      type: object
      properties:
        reply:
          type: string
        sources:
          type: array
          items:
            type: string
      required:
        - reply
        - sources

    TtsRequest:
      type: object
      properties:
        text:
          type: string
        language:
          type: string
          default: en
      required:
        - text

    TtsResponse:
      type: object
      properties:
        audioBase64:
          type: string
        mimeType:
          type: string
      required:
        - audioBase64
        - mimeType