# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Schemes Made Simple API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatResponse {
  reply: string;
  sources: string[];
}