# lib/api-zod/src/generated/types/scheme.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Schemes Made Simple API
 * OpenAPI spec version: 0.1.0
 */

export interface Scheme {
  id: number;
  name: string;
  description: string;
  category: string;
  state: string;
  ministry: string;
  launchDate: string;
  deadline?: string | null;
  benefits: string[];
  eligibility: string[];
  documents: string[];
  applicationUrl?: string | null;
  tags: string[];
}