# lib/api-zod/src/generated/types/schemeDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Schemes Made Simple API
 * OpenAPI spec version: 0.1.0
 */
import type { Scheme } from "./scheme";
import type { SchemeFaq } from "./schemeFaq";
import type { SchemeStep } from "./schemeStep";

export type SchemeDetail = Scheme & {
  steps: SchemeStep[];
  faqs: SchemeFaq[];
};