# lib/api-zod/src/generated/types/schemeListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Schemes Made Simple API
 * OpenAPI spec version: 0.1.0
 */
import type { Scheme } from "./scheme";

export interface SchemeListResponse {
  schemes: Scheme[];
  total: number;
  page: number;
  totalPages: number;
}