# lib/api-zod/src/generated/types/schemeStep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Schemes Made Simple API
 * OpenAPI spec version: 0.1.0
 */

export interface SchemeStep {
  stepNumber: number;
  title: string;
  description: string;
}