# lib/api-zod/src/generated/types/ttsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Schemes Made Simple API
 * OpenAPI spec version: 0.1.0
 */

export interface TtsResponse {
  audioBase64: string;
  mimeType: string;
}